re-export `prometheus_client` relevant types
